Only use `proxy_error_response` for proxy errors
It might be good to have an override for 500 (and even 404) errors, but
this shouldn't be the `proxy_error_response` since this is misleading
(this is specifically to avoid ugly errors when the backend can't be
contacted, or returns an ugly proxy error itself (e.g. if the backend is
an ALB).